inference_service/requirements.txt
fastapi
uvicorn
mlflow 
pydantic
requests
boto3
python-dotenv
catboost
asyncio